requirements.txt
langchain>=0.1.0
langchain-community>=0.0.2
langchain-openai>=0.0.2
openai>=1.3.7
python-dotenv>=0.19.0
tiktoken>=0.5.1
numpy>=1.24.3
pandas>=2.0.3
qtpy>=2.4.1
PyQt6>=6.6.1
PyQt6-Qt6>=6.6.1
PyQt6-sip>=13.6.0 